allow build src

+++ src/Primitives.sol
@@ -102,8 +102,8 @@ struct Route {
     uint64 gasLimit;
     uint128 baseFee;
     bytes32 gateway;
-    uint256 relativeGasPriceNumerator;
-    uint256 relativeGasPriceDenominator;
+    uint128 relativeGasPriceNumerator;
+    uint128 relativeGasPriceDenominator;
 }
 
 /**
